Update to driver and filling out logic
Added some imports to driver,
Removed unnecessary data from Ear and added melScaleSpec and melScaleCepstralCoefficients functions. Added a file to generate a dataframe for working with the guitar chord database so librosa load function doesn't need to be used every time (must be run on local machine as files are too large for github).